ad fix

+++ androidApp/features/abilities/src/main/kotlin/com/velkonost/getbetter/android/features/abilities/AbilitiesScreen.kt
@@ -78,7 +78,7 @@ fun AbilitiesScreen(
 
                     if (index % state.adPosition == 0 && index != 0) {
                         AdView(
-                            slotId = state.adId.toInt(),
+                            slotId = state.adId,
                             padding = 0
                         )
                     }
