+++ androidApp/features/addarea/src/main/kotlin/com/velkonost/getbetter/android/features/addarea/AddAreaScreen.kt
@@ -58,7 +58,7 @@ fun AddAreaScreen(
                     listState = listState,
                     isLoading = state.isLoading,
                     adPosition = state.adPosition,
-                    adSlotId = state.adId.toInt(),
+                    adSlotId = state.adId,
                     onAreaClick = {
                         scope.launch {
                             selectedAreaId.value = it
